+++ outputs.tf
@@ -5,6 +5,15 @@
 # ----------------------------------------------------------------------------------------------------------------------
 # OUTPUT ALL RESOURCES AS FULL OBJECTS
 # ----------------------------------------------------------------------------------------------------------------------
+output "google_project_iam_custom_role" {
+  description = "A map of outputs of the created google_project_iam_custom_role resources."
+  value       = google_project_iam_custom_role.roles
+}
+
+output "google_organization_iam_custom_role" {
+  value       = try(google_organization_iam_custom_role.role[0], {})
+  description = "A map of outputs of the created google_organization_iam_custom_role resource."
+}
 
 # ----------------------------------------------------------------------------------------------------------------------
 # OUTPUT MODULE CONFIGURATION