+++ test/unit-disabled/main.tf
@@ -4,23 +4,30 @@
 # The purpose is to verify no resources are created when the module is disabled.
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-variable "aws_region" {
-  description = "(Optional) The AWS region in which all resources will be created."
+variable "gcp_region" {
   type        = string
-  default     = "us-east-1"
+  description = "(Required) The gcp region in which all resources will be created."
+}
+
+variable "gcp_project" {
+  type        = string
+  description = "(Required) The ID of the project in which the resource belongs."
 }
 
 terraform {
   required_providers {
     aws = {
-      source  = "hashicorp/aws"
-      version = "~> 3.0"
+      source = "hashicorp/google"
+      # always test with exact version to catch unsupported blocks/arguments early
+      # this should match the minimal version in versions.tf
+      version = "4.8.0"
     }
   }
 }
 
-provider "aws" {
-  region = var.aws_region
+provider "google" {
+  region  = var.gcp_region
+  project = var.gcp_project
 }
 
 # DO NOT RENAME MODULE NAME
@@ -29,13 +36,11 @@ module "test" {
 
   module_enabled = false
 
-  # add all required arguments
+  role_id     = "myCustomRole"
+  title       = "My Custom Role"
+  description = "A description"
+  permissions = ["iam.roles.list", "iam.roles.create", "iam.roles.delete"]
 
-  # add all optional arguments that create additional resources
+  org_id   = "1234567"
+  projects = ["foo", "bar", "baz"]
 }
-
-# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
-# output "all" {
-#   description = "All outputs of the module."
-#   value       = module.test
-# }