+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for `null_resource` Terraform resource
+- Add support for `google_organization_iam_custom_role` Terraform resource
+- Add support for `google_project_iam_custom_role` Terraform resource
+
 
 [unreleased]: https://github.com/mineiros-io/terraform-google-iam-custom-role/compare/v0.0.1...HEAD
 <!-- [0.0.2]: https://github.com/mineiros-io/terraform-google-iam-custom-role/compare/v0.0.1...v0.0.2 -->